update#2
add a "break;" at the end of the "MFRC522_ToCard_NonBlock" function, now we can deteckt a card proberly, so when i put a false card in frot of the reader, the LED turns red, when I put a allowed one infront of it the green LED turns on.
The PROBLEM NOW:
This only works once, need to see how to let this work.